+++ src/animations/index.js
@@ -2,3 +2,4 @@ export fullCircleRotate from './fullCircleRotate';
 export pulseInfinite from './pulseInfinite';
 export slideIn from './slideIn';
 export fadeIn from './fadeIn';
+export skeleton from './skeleton';