+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Added
+
+- Add `SkeletonLoading` component.
+
 ## [3.11.0] - 2019-03-27
 
 ### Added